@alanjds 2018-10-24
The master branch is now passing.

@funny-falcon: Can you please update yours so we can have a green mark on this branch too?