@funny-falcon 2018-10-24
Not indefinitely, it is bound by O(n), where n is number of (unique) keys. And constant is usually 2, but could be a bit larger (4).
I don't think it is a real problem: dealing with returned values is usually more expensive than skipping deleted entries.
CPython doesn't "fix problem", as far as I can read and understand https://github.com/python/cpython/blob/master/Objects/dictobject.c

Ruby's hash as also tracker for first entry, because Ruby's hash has `shift` method, that returns and removes first element (while Python's `dict.popitem()` returns last element). This way Ruby's dictionary could be used as LRU structure already.
`shift` could be easily emulated for Python's dict, so if desired, then first element also could be tracked. But it doesn't speedup all iteration noticeably, it just speeds up fetching first element in some patterns, like LRU.