@funny-falcon 2018-10-24
Note that iteration over unordered dictionary still suffer from skipping empty and removed elements. Therefore there is no much difference.
Certainly, unordered dictionary need less frequent rebuilds than ordered in this pattern (delete-insert-delete-insert), but occasional rebuild still needed if key set is not stable (due to thumbstones).