@funny-falcon 2018-10-24
There could be euristic on delete: `if fill>capa/8*5 and used<capa/8 { growTable }` . I think, such euristic will fix most patological cases.