@funny-falcon 2018-10-27
Is iteration is important? I have patch that improves start of iteration at the expense of 2-4% slower putItem.

I'd prefer for faster putItem, that is why I didn't push it.